fix building tests
